pyproject.toml
[tool.poetry]
name = "utility-analysis"
version = "0.1.0"
description = "Utility cost analysis and visualization tool"
authors = ["Aaron Votre <avotre422@gmail.com>"]
packages = [{include = "scripts"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
pdfminer-six = "^20231228"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
meteostat = "^1.6.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
costs = "scripts.visualize_costs:main"
process = "scripts.process_raw_data:main"
extract = "scripts.extract_pdf_data:main"
weather = "scripts.fetch_weather:main"
efficiency = "scripts.calculate_efficiency:main"
analyze = "scripts.analyze_efficiency:main"
savings = "scripts.analyze_savings:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]